Phase C cylindrical C-grid: C.1-C.4 + accumulated WIP checkpoint
C.1 Grid + derivative + config infrastructure:
- include/core/grid_geometry.hpp: face arrays (r_face, r_face_inv, z_face)
  and a staggered flag added to GridGeometry::initialize.
- include/core/coordinate_system.hpp + src/core/infra/coordinate_system.cpp:
  StaggerType enum (Collocated, CGrid) + parser; global_stagger_type.
- include/numerics/derivatives/derivative_operators.hpp:
  StaggeredCylindricalDerivatives with one-sided face gradients,
  axis-aware flux-form divergence (2*u[0]/dr at i=0), and
  cell<->face interpolation helpers.
- src/core/runtime/runtime_config.cpp: parse grid.staggering config key.
- tests/numerics/test_staggered_derivatives.cpp.

C.2 C-grid cylindrical boundary conditions:
- src/boundary_conditions/{factory.cpp, boundary_conditions_cylindrical_cgrid.cpp}
  with rigid u[NR-1]=0 / w[NZ-1]=0, axis v=0, scalar zero-gradient ghosts,
  hydrostatic pressure extrapolation. Acoustic-substep variant scoped to
  momentum + pressure only.
- src/core/orchestration/dynamics/dynamics.cpp: BC selection through the
  factory.
- tests/dynamics/test_cylindrical_cgrid_boundary_conditions.cpp.

C.3 C-grid cylindrical initial conditions:
- src/core/orchestration/dynamics/initial_conditions_cylindrical_cgrid.cpp
  uses theta_{j+1/2} for v on the theta-face (vs collocated theta[j]).
- src/core/orchestration/dynamics/equations.cpp branches on
  (coordinate, stagger) to dispatch to the new helper.
- tests/dynamics/test_cylindrical_cgrid_initial_conditions.cpp.

C.4 Tornado C-grid dynamics scheme (CPU):
- src/dynamics/schemes/tornado/tornado_cgrid.{hpp,cpp}: axisymmetric
  TornadoCGridScheme using StaggeredCylindricalDerivatives, supercell-style
  perturbation-density buoyancy, and the standard compressible pressure
  equation. Axis ghost u[-1] = -u[0] applied inline.
- src/dynamics/factory.cpp: register tornado_cgrid + axisymmetric_cgrid alias.
- Makefile: build + test wiring (test_dynamics_tornado_cgrid).
- docs/CoordinateBackend_Plan.md: status updated, C.4 marked complete.
- tests/dynamics/test_tornado_cgrid_dynamics.cpp: 6 verification gates.

Other accumulated changes folded in for a clean tree checkpoint:
- New cartesian acoustic shader pair (acoustic_{momentum,pressure}_cartesian)
  and recompiled SPIR-V for the existing compute shaders.
- New thompson microphysics shaders (compute + SPIR-V).
- Viewer refactors: vulkan/src/{camera,core,rendering}/*, app/window/volume
  backend cleanups; new camera_input_tracker.{hpp,cpp}.
- README sweep across module dirs.
- New docs/{BUILDING,grid}.md; deletions of stale docs (FurtherOptimization,
  README, gaps, session_2026_04_20).
- Old src/advection/{advection,advection_cartesian}.cpp deleted (work moved
  to src/numerics/advection/) and src/validation/{field_contract,
  field_validation}.cpp deleted (moved to src/diagnostics/).
- scripts/setup.sh added.
- Various downstream touch-ups: dynamics.cpp, equations.cpp, headless_runtime,
  numerics.cpp, runtime_config.cpp, scheme headers + impls, time stepping,
  tests/test_harness.cpp, integration tests.

All tests pass: make test. Main binary builds clean: make -j.

+++ src/core/orchestration/physics/diffusion_step.cpp
@@ -175,24 +175,24 @@ void apply_runtime_diffusion(double dt_dynamics)
             {
                 for (int k = 0; k < NZ; ++k)
                 {
-                    float dudt = diffusion_tend_buf.dudt_diff[i][j][k];
-                    float dvdt = diffusion_tend_buf.dvdt_diff[i][j][k];
-                    float dwdt = diffusion_tend_buf.dwdt_diff[i][j][k];
-                    if (!std::isfinite(dudt)) dudt = 0.0f;
-                    if (!std::isfinite(dvdt)) dvdt = 0.0f;
-                    if (!std::isfinite(dwdt)) dwdt = 0.0f;
+                    double dudt = static_cast<double>(diffusion_tend_buf.dudt_diff[i][j][k]);
+                    double dvdt = static_cast<double>(diffusion_tend_buf.dvdt_diff[i][j][k]);
+                    double dwdt = static_cast<double>(diffusion_tend_buf.dwdt_diff[i][j][k]);
+                    if (!std::isfinite(dudt)) dudt = 0.0;
+                    if (!std::isfinite(dvdt)) dvdt = 0.0;
+                    if (!std::isfinite(dwdt)) dwdt = 0.0;
 
-                    float u_new = u[i][j][k] + dudt * static_cast<float>(dt_dynamics);
-                    float v_new = v[i][j][k] + dvdt * static_cast<float>(dt_dynamics);
-                    float w_new = w[i][j][k] + dwdt * static_cast<float>(dt_dynamics);
+                    double u_new = static_cast<double>(u[i][j][k]) + dudt * dt_dynamics;
+                    double v_new = static_cast<double>(v[i][j][k]) + dvdt * dt_dynamics;
+                    double w_new = static_cast<double>(w[i][j][k]) + dwdt * dt_dynamics;
 
-                    if (!std::isfinite(u_new)) u_new = 0.0f;
-                    if (!std::isfinite(v_new)) v_new = 0.0f;
-                    if (!std::isfinite(w_new)) w_new = 0.0f;
+                    if (!std::isfinite(u_new)) u_new = 0.0;
+                    if (!std::isfinite(v_new)) v_new = 0.0;
+                    if (!std::isfinite(w_new)) w_new = 0.0;
 
-                    u[i][j][k] = clamp_wind_horizontal_ms(u_new);
-                    v[i][j][k] = clamp_wind_horizontal_ms(v_new);
-                    w[i][j][k] = clamp_wind_vertical_ms(w_new);
+                    u[i][j][k] = clamp_wind_horizontal_ms(static_cast<float>(u_new));
+                    v[i][j][k] = clamp_wind_horizontal_ms(static_cast<float>(v_new));
+                    w[i][j][k] = clamp_wind_vertical_ms(static_cast<float>(w_new));
                 }
             }
         }
@@ -207,19 +207,19 @@ void apply_runtime_diffusion(double dt_dynamics)
             {
                 for (int k = 0; k < NZ; ++k)
                 {
-                    float dthetadt = diffusion_tend_buf.dthetadt_diff[i][j][k];
-                    float dqvdt = diffusion_tend_buf.dqvdt_diff[i][j][k];
-                    if (!std::isfinite(dthetadt)) dthetadt = 0.0f;
-                    if (!std::isfinite(dqvdt)) dqvdt = 0.0f;
+                    double dthetadt = static_cast<double>(diffusion_tend_buf.dthetadt_diff[i][j][k]);
+                    double dqvdt = static_cast<double>(diffusion_tend_buf.dqvdt_diff[i][j][k]);
+                    if (!std::isfinite(dthetadt)) dthetadt = 0.0;
+                    if (!std::isfinite(dqvdt)) dqvdt = 0.0;
 
-                    float theta_new = theta[i][j][k] + dthetadt * static_cast<float>(dt_dynamics);
-                    float qv_new = qv[i][j][k] + dqvdt * static_cast<float>(dt_dynamics);
+                    double theta_new = static_cast<double>(theta[i][j][k]) + dthetadt * dt_dynamics;
+                    double qv_new = static_cast<double>(qv[i][j][k]) + dqvdt * dt_dynamics;
 
-                    if (!std::isfinite(theta_new)) theta_new = static_cast<float>(theta0);
-                    if (!std::isfinite(qv_new)) qv_new = 0.0f;
+                    if (!std::isfinite(theta_new)) theta_new = theta0;
+                    if (!std::isfinite(qv_new)) qv_new = 0.0;
 
-                    theta[i][j][k] = clamp_theta_k(theta_new);
-                    qv[i][j][k] = clamp_qv_kgkg(qv_new);
+                    theta[i][j][k] = clamp_theta_k(static_cast<float>(theta_new));
+                    qv[i][j][k] = clamp_qv_kgkg(static_cast<float>(qv_new));
                 }
             }
         }
